requirements.txt
python-nmap
subprocess
argparse